requirements.txt
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-telegram-bot>=20.6
apscheduler>=3.10.0
ta>=0.11.0
plotly>=5.18.0
httpx>=0.25.0
pydantic>=2.5.0
jinja2>=3.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
